+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-03T19:08:20.176Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-04T06:09:33.474Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")