+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K48FP0GHZQXR86Y0X12QMTMV"
+        "01K49NGQXKKXEJNW6B3Y4HAN5H"
     )
     r"""The id of the resource"""