+++ packages/orq-rc/src/orq_ai_sdk/models/retrievecontactop.py
@@ -101,7 +101,7 @@ class RetrieveContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-03T19:08:20.176Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-04T06:09:33.474Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")