+++ packages/orq-rc/RELEASES.md
@@ -2599,3 +2599,13 @@ Based on:
 - [python v3.12.0-rc.36] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.36] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.36 - packages/orq-rc
+
+## 2025-09-04 06:11:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.609.0 (2.692.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.37] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.37] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.37 - packages/orq-rc